Moved DataAccumulator to a separate file
